+++ lib/src/version-specific.ts
@@ -1,10 +1,16 @@
 import {
   CommandNode,
   CommandNodeChildren,
-  CommandNodeType
+  CommandNodeType,
+  CommandNodeContext
 } from "./command-manifest";
 import { ExtendedLanguageGrammar, GrammarNode } from "./language-grammar";
 import { deepCopy } from "./utils";
+import { getGrammarNodeGenerator } from "./version-specific/grammar-node-generators";
+
+export const ERROR_PATTERNS = {
+  patterns: [{ include: "#error.command_line" }]
+};
 
 export function getCommandNodeChildNames(commandNode: CommandNode): string[] {
   return Object.keys(commandNode.children ? commandNode.children : []);
@@ -18,23 +24,54 @@ export function getGrammarGroupName(...components: string[]): string {
   return "generated.commands." + components.join(".");
 }
 
-export function makeGrammarNode(
-  commandNode: CommandNode,
-  name: string,
-  breadcrumb: string[]
+export function makeLiteralGrammarNode(
+  context: CommandNodeContext
 ): GrammarNode {
-  // FIXME deal with node type and parser
-  const grammarGroupName = getGrammarGroupName(...breadcrumb, name);
+  const grammarGroupName = getGrammarGroupName(
+    ...context.breadcrumb,
+    context.name
+  );
   return {
-    begin: commandNode.executable
-      ? `(${name}){{cbx_true}}`
-      : `(${name}){{cbx_false}}`,
+    begin: context.node.executable
+      ? `(${context.name}){{cbx_true}}`
+      : `(${context.name}){{cbx_false}}`,
     end: "{{ln}}",
     beginCaptures: {
-      1: { name: breadcrumb.length > 0 ? "#subcommand" : "#command" }
+      1: { name: context.breadcrumb.length > 0 ? "#subcommand" : "#command" }
     },
     patterns: [{ include: `#${grammarGroupName}` }]
   } as GrammarNode;
+}
+
+function printError(context: CommandNodeContext, message: string) {
+  const nodePath = [...context.breadcrumb, context.name].join(".");
+  console.error(`${message} at node ${nodePath}`);
+}
+
+export function makeArgumentGrammarNode(
+  context: CommandNodeContext
+): GrammarNode {
+  const nodeParserGenerator = getGrammarNodeGenerator(context.node.parser);
+
+  if (nodeParserGenerator) {
+    return nodeParserGenerator.generate(context);
+  }
+
+  printError(context, `Unknown parser "${context.node.parser}"`);
+  return ERROR_PATTERNS;
+}
+
+export function makeGrammarNode(context: CommandNodeContext): GrammarNode {
+  if (context.node.type === CommandNodeType.LITERAL) {
+    return makeLiteralGrammarNode(context);
+  }
+
+  if (context.node.type === CommandNodeType.ARGUMENT) {
+    return makeArgumentGrammarNode(context);
+  }
+
+  printError(context, `Invalid node type "${context.node.type}"`);
+  return ERROR_PATTERNS;
 }
 
 export function makeGrammarGroup(
@@ -57,27 +94,30 @@ export function makeGrammarGroup(
 
 export function augmentNode(
   grammar: ExtendedLanguageGrammar,
-  commandNode: CommandNode,
-  name: string,
-  breadcrumb: string[]
+  context: CommandNodeContext
 ) {
-  const nextBreadcrumb = [...breadcrumb, name];
+  const nextBreadcrumb = [...context.breadcrumb, context.name];
   const grammarNodeName = getGrammarNodeName(...nextBreadcrumb);
   const grammarGroupName = getGrammarGroupName(...nextBreadcrumb);
 
   // register a grammar pattern for the command node
-  const grammarNode = makeGrammarNode(commandNode, name, breadcrumb);
+  const grammarNode = makeGrammarNode(context);
   grammar.repository[grammarNodeName] = grammarNode;
 
   // register a grammar pattern with includes for all children command nodes
   // we do this separately to reduce nested patterns and to trivialize redirects
-  const grammarGroup = makeGrammarGroup(commandNode, nextBreadcrumb);
+  const grammarGroup = makeGrammarGroup(context.node, nextBreadcrumb);
   grammar.repository[grammarGroupName] = grammarGroup;
 
   // recursively add children to the grammar
-  getCommandNodeChildNames(commandNode).forEach(childName => {
-    const childCommandNode = commandNode.children[childName];
-    augmentNode(grammar, childCommandNode, childName, nextBreadcrumb);
+  getCommandNodeChildNames(context.node).forEach(childName => {
+    const childCommandNode = context.node.children[childName];
+    const childContext = new CommandNodeContext(
+      childCommandNode,
+      childName,
+      nextBreadcrumb
+    );
+    augmentNode(grammar, childContext);
   });
 }
 
@@ -88,11 +128,11 @@ export function augmentGrammar(
   // make a copy of the given grammar
   const grammar = deepCopy(baseGrammar) as ExtendedLanguageGrammar;
 
-  // DELETE
+  // DELETEME
   commands = {
     type: CommandNodeType.ROOT,
     children: {
-      advancement: commands.children["advancement"]
+      effect: commands.children["effect"]
     }
   };
 
@@ -100,7 +140,7 @@ export function augmentGrammar(
   const grammarRootGroup = makeGrammarGroup(commands, []);
   grammar.repository["generated.commands"] = grammarRootGroup;
 
-  // DELETE
+  // DELETEME
   grammarRootGroup.patterns = [
     { include: "#generated.command.say" },
     { include: "#generated.command.tellraw" },
@@ -111,7 +151,12 @@ export function augmentGrammar(
   // recursively add top-level commands to the grammar
   getCommandNodeChildNames(commands).forEach(childName => {
     const childCommandNode = commands.children[childName];
-    augmentNode(grammar, childCommandNode, childName, []);
+    const childContext = new CommandNodeContext(
+      childCommandNode,
+      childName,
+      []
+    );
+    augmentNode(grammar, childContext);
   });
 
   return grammar;