+++ lib/src/version-specific/grammar-node-generators/entity-grammar-node-generator.ts
@@ -1,0 +1,10 @@
+import { CommandNodeContext } from "../../command-manifest";
+import { GrammarNode } from "../../language-grammar";
+import { GrammarNodeGenerator } from "./abc/grammar-node-generator";
+
+export class EntityGrammarNodeGenerator extends GrammarNodeGenerator {
+  generate(context: CommandNodeContext): GrammarNode {
+    // FIXME implement
+    return { patterns: [] };
+  }
+}