+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- Improved architecture for version-specific grammar generator
+
 ## [0.15.0] - 2019-09-20
 ### Added
 - Implemented special case to support greedy commands (say), as per #25